An unstated enum tag is the previous tag plus one; a collision is an error.
A bare first member is 0 (a flag, 1); an unstated member takes the previous
tag's successor, in a flags enum the bit above it; a stated member takes its
literal, negated or not. One kernel, `memberTags`, computes them for both
enum shapes. Registration diagnoses a tag that is not a literal where it is
written, and two members reaching one tag at the later member with a note at
the holder. The IR carries the tag as i64 so a negative tag flows through
construction, `match`, `@tag`, reflection and the formatter at every width.

+++ src/ir/inst.zig
@@ -463,7 +463,7 @@ pub const Aggregate = struct {
 };
 
 pub const EnumInit = struct {
-    tag: u32,
+    tag: i64,
     payload: Ref, // Ref.none if no payload
 };
